Move jx_eval() to a new module.
