+++ dttools/src/Makefile
@@ -59,6 +59,7 @@ SOURCES = \
 	jx_print.c \
 	jx_table.c \
 	jx_export.c \
+	jx_eval.c \
 	link.c \
 	link_auth.c \
 	list.c \
